do all exercises in 4 expect last

+++ osa4/blogilista/controllers/blogs.js
@@ -1,15 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const GetTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
 blogsRouter.get('/', async(request, response) => {
   const blogs = await Blog
@@ -19,11 +11,10 @@ blogsRouter.get('/', async(request, response) => {
   
 blogsRouter.post('/', async(request, response) => {
   try {
-    const decodedToken = jwt.verify(GetTokenFrom(request), process.env.SECRET)
-    if (!decodedToken.id) {
+    const user = request.user
+    if (!user) {
       return response.status(401).json({ error: 'token invalid' })
     }
-    const user = await User.findById(decodedToken.id)
     const blog = new Blog({
       ...request.body,
       user: user._id
@@ -35,8 +26,17 @@ blogsRouter.post('/', async(request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', async(request, response) => {
   try {
+    const user = request.user
+    if (!user) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== user._id.toString()) {
+      return response.status(401).json({ error: 'unauthorized' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
   }
@@ -57,3 +57,4 @@ blogsRouter.put('/:id', async (request, response) => {
 
 
 module.exports = blogsRouter
+
